:bug: Fix router not finish loading request handler file before accepting request
Currently, this module return a router for user to use in the server init file.
But express router does not handle async function, so the async import statement
are not being waited, leading to the problem of router accepting request before
properly loading all the handler file.
To deal with this, change the router into async middleware seems to solve the
before mentioned problem.
Also, this commit deal with the forEach problem, in which the sort order of path
is changed during import file operation, as the import ops completions dont follow
any order.